Game Mode Manager
Functionality is in to switch between game modes, play them, and stop them. Went with an interface for the game modes so we can expand as needed easily, and have a single manager in scene to handle it. Currently am just using 1,2,  and 3 to switch between Timed Mode, Numbered Mode and Moving mode respectively, as well as 9 to start and 0 to stop the game. This is just game modes, nothing with the targets has been coded.